Refactor to top-level sl/rl division
