+++ loss_landscapes/common/model_interface/model_agnostic_factories.py
@@ -1,7 +1,7 @@
-import loss_landscapes.model_interface.model_wrapper as model_wrapper
-import loss_landscapes.model_interface.model_tensor as model_tensor
-import loss_landscapes.model_interface.torch.torch_wrappers as torch_wrappers
-import loss_landscapes.model_interface.torch.torch_tensor as torch_tensor
+import loss_landscapes.common.model_interface.model_wrapper as model_wrapper
+import loss_landscapes.common.model_interface.model_tensor as model_tensor
+import loss_landscapes.common.model_interface.torch.torch_wrappers as torch_wrappers
+import loss_landscapes.common.model_interface.torch.torch_tensor as torch_tensor
 
 
 def rand_u_like(example_tensor: model_tensor.ParameterTensor) -> model_tensor.ParameterTensor:
