+++ loss_landscapes/common/model_interface/torch/torch_vector.py
@@ -1,4 +1,4 @@
-import loss_landscapes.model_interface.model_vector as model_vector
+import common.model_interface.model_vector as model_vector
 
 
 class TorchParameterVector(model_vector.ParameterVector):
