--- loss_landscapes/connecting_paths/__init__.py
@@ -1,1 +1,0 @@
-from loss_landscapes.connecting_paths.connecting_paths import auto_neb, garipov_curve_search
