--- loss_landscapes/evaluators/__init__.py
@@ -1,1 +1,0 @@
-from loss_landscapes.evaluators.evaluator import Evaluator, EvaluatorPipeline
