--- loss_landscapes/evaluators/torch/__init__.py
@@ -1,2 +1,0 @@
-from loss_landscapes.evaluators.torch.sl_evaluators import *
-from loss_landscapes.evaluators.torch.metric_evaluators import *
