+++ loss_landscapes/rl/evaluators/torch/rl_evaluators.py
@@ -7,13 +7,7 @@ of reinforcement learning models in RL environments.
 
 
 from abc import ABC, abstractmethod
-import numpy as np
-import torch
-import torch.autograd
-from loss_landscapes.evaluators.evaluator import Evaluator
-from loss_landscapes.model_interface.model_agnostic_factories import wrap_model
-from loss_landscapes.model_interface.torch.torch_wrappers import TorchNamedParameterWrapper
-from loss_landscapes.model_interface.torch.torch_tensor import rand_u_like
+from loss_landscapes.common.evaluators.evaluator import Evaluator
 
 
 class GymEvaluator(Evaluator, ABC):
