--- loss_landscapes/saddle_points/__init__.py
@@ -1,1 +1,0 @@
-from loss_landscapes.saddle_points.saddle_points import find_saddle_point
