--- loss_landscapes/saddle_points/saddle_points.py
@@ -1,10 +1,0 @@
-""" Functions for exploring and analyzing the saddle points of a loss landscape. """
-
-
-import numpy as np
-
-
-def find_saddle_point() -> np.ndarray:
-    """ Gradient traversal to find saddle points, inspired by https://arxiv.org/abs/1406.2572 """
-    # todo
-    raise NotImplementedError()
