+++ loss_landscapes/sl/evaluators/torch/evaluators.py
@@ -11,10 +11,9 @@ from abc import ABC, abstractmethod
 import numpy as np
 import torch
 import torch.autograd
-from loss_landscapes.evaluators.evaluator import Evaluator
-from loss_landscapes.model_interface.model_agnostic_factories import wrap_model
-from loss_landscapes.model_interface.torch.torch_wrappers import TorchNamedParameterWrapper
-from loss_landscapes.model_interface.torch.torch_tensor import rand_u_like
+from loss_landscapes.common.evaluators.evaluator import Evaluator
+from loss_landscapes.common.model_interface.model_agnostic_factories import wrap_model, rand_u_like
+from loss_landscapes.common.model_interface.torch.torch_wrappers import TorchNamedParameterWrapper
 
 
 class TorchSupervisedEvaluator(Evaluator, ABC):
