--- loss_landscapes/trajectories/__init__.py
@@ -1,1 +1,0 @@
-from loss_landscapes.trajectories.trackers import ProjectingTrajectoryTracker
