--- loss_landscapes/utils/__init__.py
@@ -1,1 +1,0 @@
-from loss_landscapes.utils.copy import deepcopy_model
